+++ .jest-skip/validate-coderabbit-yml.test.js
@@ -1,57 +1,34 @@
-// validate-coderabbit-yml.test.js
-// Jest test for validate-coderabbit-yml.cjs
-
-const { execSync } = require("child_process");
-const fs = require("fs");
+/**
+ * @jest-environment jsdom
+ */
 const path = require("path");
+const { execSync } = require("child_process");
 
-describe("validate-coderabbit-yml.cjs", () => {
-  const scriptPath = path.resolve(__dirname, "../validate-coderabbit-yml.cjs");
-  const ymlPath = path.resolve(__dirname, "../../../.coderabbit.yml");
-  const backupPath = ymlPath + ".bak";
-
-  beforeAll(() => {
-    // Backup the original .coderabbit.yml if it exists
-    if (fs.existsSync(ymlPath)) {
-      fs.copyFileSync(ymlPath, backupPath);
-    }
-    // Write a minimal valid .coderabbit.yml
-    fs.writeFileSync(
-      ymlPath,
-      'reviews:\n  path_filters: ["src/"]\n  auto_review: true\n',
+describe("Coderabbit YML Validation", () => {
+  it("should validate a correct coderabbit.yml file", () => {
+    const script = path.resolve(__dirname, "../validate-coderabbit-yml.cjs");
+    const file = path.resolve(
+      __dirname,
+      "../__fixtures__/valid-coderabbit.yml",
     );
+    const result = execSync(`node ${script} ${file}`, { encoding: "utf8" });
+    expect(result).toMatch(/\.coderabbit\.yml is valid!/i);
   });
 
-  afterAll(() => {
-    // Restore the original .coderabbit.yml
-    if (fs.existsSync(backupPath)) {
-      fs.copyFileSync(backupPath, ymlPath);
-      fs.unlinkSync(backupPath);
-    } else {
-      fs.unlinkSync(ymlPath);
-    }
-  });
-
-  it("validates a correct .coderabbit.yml and exits 0", () => {
-    let output = "";
-    expect(() => {
-      output = execSync(`node ${scriptPath}`, { encoding: "utf8" });
-    }).not.toThrow();
-    expect(output).toMatch(/\.coderabbit\.yml is valid!/);
-  });
-
-  it("fails if required field is missing", () => {
-    // Write an invalid .coderabbit.yml (missing reviews)
-    fs.writeFileSync(ymlPath, "notreviews: true\n");
+  it("should fail on an invalid coderabbit.yml file", () => {
+    const script = path.resolve(__dirname, "../validate-coderabbit-yml.cjs");
+    const file = path.resolve(
+      __dirname,
+      "../__fixtures__/invalid-coderabbit.yml",
+    );
     let error = null;
     try {
-      execSync(`node ${scriptPath}`, { encoding: "utf8", stdio: "pipe" });
+      execSync(`node ${script} ${file}`, { encoding: "utf8" });
     } catch (e) {
       error = e;
     }
     expect(error).toBeTruthy();
-    expect(error.stdout || error.message).toMatch(
-      /Missing required top-level field: reviews/,
-    );
+    expect(error.stdout).toMatch(/Invalid \.coderabbit\.yml/i);
+    expect(error.stderr).toMatch(/Missing required top-level field: reviews/);
   });
 });
